Refactor transcoder.transcoder -> transcoder
