+++ core/auto_process/movies.py
@@ -7,10 +7,9 @@ import time
 import requests
 
 import core
-from core import logger
+from core import logger, transcoder
 from core.nzbToMediaSceneExceptions import process_all_exceptions
 from core.nzbToMediaUtil import convert_to_ascii, find_download, find_imdbid, import_subs, list_media_files, remote_dir, report_nzb, remove_dir, server_responding
-from core.transcoder import transcoder
 
 requests.packages.urllib3.disable_warnings()
