+++ core/auto_process/tv.py
@@ -9,11 +9,10 @@ import time
 import requests
 
 import core
-from core import logger
+from core import logger, transcoder
 from core.nzbToMediaAutoFork import auto_fork
 from core.nzbToMediaSceneExceptions import process_all_exceptions
 from core.nzbToMediaUtil import convert_to_ascii, flatten, import_subs, list_media_files, remote_dir, report_nzb, remove_dir, server_responding
-from core.transcoder import transcoder
 
 requests.packages.urllib3.disable_warnings()
