--- core/transcoder/__init__.py
@@ -1,2 +1,0 @@
-# coding=utf-8
-__author__ = 'Justin'
